refactor(game): move read-only queries to data-access/reads
Now that lazy-regen is decoupled (data-access/tile-regen), the read getters
extract cleanly. Moves the contiguous read cluster — getPlayerServer,
getOwnedTilesServer, getOwnedMapTilesServer, getAllMapTilesServer,
getMapTilesInBoundsServer, getMyMapServer, getAllOwnerSummariesServer,
getTileServer — plus the OwnerSummary/MyMapView result types and the
VIEWPORT_TILE_LIMIT cap into lib/game/data-access/reads.ts.

These are dependency-injected reads (adminDbOrThrow) with no transaction state;
reads.ts depends only on ../turns, ../world-gen, ../types, and
data-access/{collections,db,tile-regen} — no data-server import, no cycle.
data-server imports the getters back (5 are still called internally) and
re-exports all eight + both types, so `@/lib/game/data-server` import sites are
unaffected. `applyLazyRegen`/`applyLazyRegenBatch` and the `MapTile` type are no
longer referenced in data-server and were dropped from its imports.

Behavior-preserving — code moved verbatim. data-server.ts 7,337 → 7,040
(7,865 → 7,040 across the full Phase-4 chain — ~825 lines / 10% off the
god-object, now organized under lib/game/data-access/).

Verified: game + game-API suites 1780/1780, tsc + eslint clean.

Stacked on #1572.
